// backend/src/models/SaleRecord.d.ts
export interface SaleRecord {
    transactionId: number;
    date: string;
    customerId: string;
    customerName: string;
    phoneNumber: string;
    gender: string;
    age: number;
    customerRegion: string;
    customerType: string;
    productId: string;
    productName: string;
    brand: string;
    productCategory: string;
    tags: string;
    quantity: number;
    pricePerUnit: number;
    discountPercentage: number;
    totalAmount: number;
    finalAmount: number;
    paymentMethod: string;
    orderStatus: string;
    deliveryType: string;
    storeId: string;
    storeLocation: string;
    salespersonId: string;
    employeeName: string;
}
//# sourceMappingURL=SaleRecord.d.ts.map